extracting construct local function into its own file
relates to #79
