+++ src/helper_functions/constructLocal.js
@@ -1,0 +1,14 @@
+
+
+function constructLocal(senderID, key, answerObjects) {
+  const messageData = {
+    recipient: {
+      id: senderID,
+    },
+    message: answerObjects[key],
+  };
+
+  return messageData;
+}
+
+module.exports = constructLocal;
